requirements.txt
flask==2.2.3
werkzeug==2.2.3
PyPDF2==3.0.1
Pillow==9.5.0
python-magic==0.4.27
zlib==1.2.13
compress-pickle==2.1.0
base64==1.0.0
